Assignment 1 Brandon Juffe
For the First Question I added 3 + 4 from both Register 1 and 2 with the result added into three which is 7
For the Second Question, my first try was a success with subtracting register 2 (Value of 1) from register 1 (Value of 3) for the answer of 2 in register 4.
For the Failure on question 2 I tried subtracting a value that was bigger than the value being subtracted from which made the program reguster a 0 value in register 4 which is incorrect.